Update checkers in Gerrit after deployment
Add a playbook which updates the configuration of checkers in Gerrit.
It accepts a list of dictionaries as input, each corresponding to
a checker.  If the UUID of the checker does not already exist in
Gerrit, it is created.  If it does exist, it is updated.  Fields
which are not present in the input are not touched (nor are they
considered when comparing the declared state to the existing state).

Change-Id: I3e70b15717aa1ff89cdcf5dc4dc52187aa0347aa