Limit Nodepool to 4 servers
We have a 72vcpu quota, and nodepool does not have support for
automatically detecting the quota in GCE yet.  If we request too
many, nodepool will return an error.  Until we improve the GCE
driver in nodepool to handle this automatically, limit our usage
to 4 servers, which should stay under the quota since our largest
usage is 16vcpus.

Change-Id: Ic7d08593889891ded4aec8114594c068f5330f42