Use TLS Zookeeper connections
This creates 2 TLS secrets: one for internal use among Zookeeper quorum
members, and a second for use by Zuul and Nodepool as clients.

It uses pre-generated output from the helm charts (with a locally applied
patch to add TLS support while we wait for an upstream PR which adds it
to merge).  These switch the existing cluster to use server-size and
client-side TLS.

Zuul and Nodepool configurations are also updated to use the new client
certs.

Change-Id: I6951c39501c61291d2f7eec43426251c603bc537